create custom modal component and organize files
