+++ package.json
@@ -10,11 +10,13 @@
   },
   "dependencies": {
     "axios": "^1.3.4",
+    "formik": "^2.2.9",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-icons": "^4.8.0",
     "react-query": "^3.39.3",
     "react-router-dom": "^6.10.0",
+    "yup": "^1.0.2",
     "zustand": "^4.3.7"
   },
   "devDependencies": {
